Move JS code for state machine to @expo/use-updates
